Fixed crash in :Map label
A crash was occurring when a user would enter into map selection without first assigning a game mode. The reason being was because I added an earlier feature to not display maps that were incompatible with TDM if that game mode was currently selected, so I added a work around.